# AERIS/backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AERIS Backend",
    "description": "Maryland statewide site screening, scoped parcel investigation, mapped physical constraints, public grid context, and planning context.",
    "version": "0.6.1"
  },
  "paths": {
...
    "/analysis/statewide/grid": {
      "get": {
        "tags": [
          "statewide"
        ],
        "summary": "Statewide Grid",
        "operationId": "statewide_grid_analysis_statewide_grid_get",
        "parameters": [
          {
            "name": "score_type",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "technical",
                "effective"
              ],
              "type": "string",
              "default": "technical",
              "title": "Score Type"
            }
          },
          {
            "name": "minimum_score",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "maximum": 1.0,
              "minimum": 0.0,
              "default": 0.0,
              "title": "Minimum Score"
            }
          },
          {
            "name": "maximum_score",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "maximum": 1.0,
              "minimum": 0.0,
              "default": 1.0,
              "title": "Maximum Score"
            }
          },
          {
            "name": "eligibility",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "all",
                "auto",
                "exploration"
              ],
              "type": "string",
              "default": "all",
              "title": "Eligibility"
            }
          },
          {
            "name": "equity_gate",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Equity Gate"
            }
          },
          {
            "name": "excluded",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Excluded"
            }
          },
          {
            "name": "county",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "County"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 30000,
              "minimum": 1,
              "default": 30000,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeoJSONFeatureCollection"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/analysis/statewide/candidate-zones": {
      "get": {
        "tags": [
          "statewide"
        ],
        "summary": "Statewide Candidate Zones",
        "operationId": "statewide_candidate_zones_analysis_statewide_candidate_zones_get",
        "parameters": [
          {
            "name": "mode",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "top",
                "auto",
                "exploration"
              ],
              "type": "string",
              "default": "top",
              "title": "Mode"
            }
          },
          {
            "name": "minimum_score",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "maximum": 1.0,
              "minimum": 0.0,
              "default": 0.0,
              "title": "Minimum Score"
            }
          },
          {
            "name": "maximum_score",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "maximum": 1.0,
              "minimum": 0.0,
              "default": 1.0,
              "title": "Maximum Score"
            }
          },
          {
            "name": "county",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "County"
            }
          },
          {
            "name": "top_n",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 100,
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "title": "Top N"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeoJSONFeatureCollection"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/analysis/statewide/cells/{cell_id}": {
      "get": {
        "tags": [
          "statewide"
        ],
        "summary": "Statewide Cell Detail",
        "operationId": "statewide_cell_detail_analysis_statewide_cells__cell_id__get",
        "parameters": [
          {
            "name": "cell_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Cell Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatewideCellDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/analysis/statewide/zones/{zone_id}": {
      "get": {
        "tags": [
          "statewide"
        ],
        "summary": "Statewide Zone Detail",
        "operationId": "statewide_zone_detail_analysis_statewide_zones__zone_id__get",
        "parameters": [
          {
            "name": "zone_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Zone Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatewideZoneDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "GeoJSONFeatureCollection": {
        "properties": {
          "type": {
            "type": "string",
            "const": "FeatureCollection",
            "title": "Type",
            "default": "FeatureCollection"
          },
          "features": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Features"
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "features"
        ],
        "title": "GeoJSONFeatureCollection"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "StatewideCellDetailResponse": {
        "properties": {
          "cell_id": {
            "type": "string",
            "title": "Cell Id"
          },
          "location": {
            "$ref": "#/components/schemas/StatewideLocation"
          },
          "decision": {
            "$ref": "#/components/schemas/StatewideDecision"
          },
          "scores": {
            "$ref": "#/components/schemas/StatewideScores"
          },
          "evidence": {
            "additionalProperties": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "object",
            "title": "Evidence"
          },
          "community_impact": {
            "additionalProperties": true,
            "type": "object",
            "title": "Community Impact"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "cell_id",
          "location",
          "decision",
          "scores"
        ],
        "title": "StatewideCellDetailResponse"
      },
...
      "StatewideDecision": {
        "properties": {
          "model_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Status"
          },
          "hard_excluded": {
            "type": "boolean",
            "title": "Hard Excluded"
          },
          "exclusion_reasons": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Exclusion Reasons"
          },
          "auto_screen_eligible": {
            "type": "boolean",
            "title": "Auto Screen Eligible"
          },
          "exploration_screen_eligible": {
            "type": "boolean",
            "title": "Exploration Screen Eligible"
          },
          "automated_recommendation_ready": {
            "type": "boolean",
            "title": "Automated Recommendation Ready"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "hard_excluded",
          "auto_screen_eligible",
          "exploration_screen_eligible",
          "automated_recommendation_ready"
        ],
        "title": "StatewideDecision"
      },
...
      "StatewideLocation": {
        "properties": {
          "latitude": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latitude"
          },
          "longitude": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Longitude"
          },
          "county_fips": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "County Fips"
          },
          "county_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "County Name"
          },
          "tract_geoid": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tract Geoid"
          },
          "land_fraction": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Land Fraction"
          },
          "cell_area_sq_km": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cell Area Sq Km"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "title": "StatewideLocation"
      },
...
      "StatewideScores": {
        "properties": {
          "technical_suitability": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Technical Suitability"
          },
          "effective_suitability": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Effective Suitability"
          },
          "criteria": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ]
            },
            "type": "object",
            "title": "Criteria"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "title": "StatewideScores"
      },
...
      "StatewideZoneDetailResponse": {
        "properties": {
          "zone_id": {
            "type": "string",
            "title": "Zone Id"
          },
          "mode": {
            "type": "string",
            "title": "Mode"
          },
          "properties": {
            "additionalProperties": true,
            "type": "object",
            "title": "Properties"
          },
          "membership": {
            "$ref": "#/components/schemas/ZoneMembership"
          },
          "member_summary": {
            "$ref": "#/components/schemas/ZoneMemberSummary"
          },
          "interpretation": {
            "type": "string",
            "title": "Interpretation"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "zone_id",
          "mode",
          "membership",
          "member_summary",
          "interpretation"
        ],
        "title": "StatewideZoneDetailResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "ZoneMemberSummary": {
        "properties": {
          "technical_score": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ]
            },
            "type": "object",
            "title": "Technical Score"
          },
          "equity_gate_counts": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Equity Gate Counts"
          },
          "county_counts": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "County Counts"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "title": "ZoneMemberSummary"
      },
      "ZoneMembership": {
        "properties": {
          "member_count": {
            "type": "integer",
            "title": "Member Count"
          },
          "member_cell_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Member Cell Ids"
          },
          "member_cell_ids_truncated": {
            "type": "boolean",
            "title": "Member Cell Ids Truncated"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "member_count",
          "member_cell_ids_truncated"
        ],
        "title": "ZoneMembership"
      }
    }
  }
}